requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.37
asyncpg
psycopg2-binary
alembic==1.14.1
fastapi-users[sqlalchemy]
pydantic==2.10.6
python-dotenv==1.0.1
redis==5.2.1
fastapi-cache2[redis]
gunicorn
python-multipart
